% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gss_1.R
\docType{data}
\name{Gss_1}
\alias{Gss_1}
\title{Population of various Religious Groups in Ghana from the Ghana Population
and Housing Census in 2021}
\format{
A data frame with 160 rows and 4 variables:
\describe{
\item{Religious_Affiliation}{the type of religion a certain group of people
are associated with(Christian, Islam,Traditionalist(local Ghanaian religion),
No Religion, Other Religion(apart from the ones stated))}
\item{Locality}{The type of area a particular group of people are located
(Rural, Urban)}
\item{Region}{All 16 Regions of Ghana (see...\url{https://www.ghanamissionun.org/map-regions-in-ghana/}
for more details on the 16 Regions in Ghana)}
\item{Population}{The number of people belonging to a particular Religious_Affiliation}
}
}
\source{
\url{https://statsbank.statsghana.gov.gh/pxweb/en/PHC\%202021\%20StatsBank/}
}
\usage{
Gss_1
}
\description{
A dataset containing the number of people in all the 16 Regions of Ghana and
their various Religions Affiliations.The variables are as follows:
}
\keyword{datasets}
